Use UnityVersion, NuGetPackageVersion
