+++ src/NuGetForUnity/Editor/Models/UnityVersion.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 294ab99fe7354275bb1500695b816801
+timeCreated: 1707812879
